backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
daphne==4.0.0
channels==4.0.0
twilio==9.0.1
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.0
wheel
setuptools==65.5.0